Update instance put endpoint to return final document in response
