+++ models/dataset.go
@@ -223,29 +223,6 @@ type VersionLinks struct {
 	Version    *LinkObject `bson:"version,omitempty"     json:"-"`
 }
 
-var validVersionStates = map[string]int{
-	EditionConfirmedState: 1,
-	AssociatedState:       1,
-	PublishedState:        1,
-}
-
-// CheckState checks state against a whitelist of valid states
-func CheckState(docType, state string) error {
-	var states map[string]int
-	switch docType {
-	case "version":
-		states = validVersionStates
-	default:
-		states = validStates
-	}
-
-	if states[state] == 1 {
-		return nil
-	}
-
-	return errs.ErrResourceState
-}
-
 // CreateDataset manages the creation of a dataset from a reader
 func CreateDataset(reader io.Reader) (*Dataset, error) {
 	b, err := ioutil.ReadAll(reader)
